Saving inventory & player position
I had to remove the old inventory system (or modify it) in order to create a saving system for the inventory.

I created some scripts (mostly following a tutorial) for saving inventory slots, items and player position.

Next time i think i will reimplement picking up stuff and them updating in the inventory, but also saving which scene we were last time we closed the game.