Fixed some collision errors and some inventory minor inventory fixes
Fixed a bug where the player tries gets a component from the object it collides with, even if the object isn't a interactable item. Now checks if the interact handler  is a component of said object.

The inventory on first load displayed an empty item information panel. Now it displays an actual empty panel (no white image or random name or description). Everytime the player closes the inventory, all information objects get hidden, until the player clicks on one of the items in the inventory.